delete for view all post is working fine and added some js in it

+++ admin/includess/view_all_users.php
@@ -65,7 +65,7 @@
         echo "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
         echo "<td><a href='users.php?change_to_subscriber={$user_id}'>Subscriber</a></td>";
         echo "<td><a href='users.php?source=edit_user&p_id={$user_id}'>Edit</a></td>";
-        echo "<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
+        echo "<td><a onclick=\"javascript: return confirm('Are you sure you want to delete this post?');\" href='users.php?delete={$user_id}'>Delete</a></td>";
 //        echo "</tr>";
 
     }
@@ -107,7 +107,7 @@ if(isset($_GET['change_to_subscriber'])){
 
     if (isset($_GET['delete'])) {
         if (isset($_SESSION['user_role'])) {
-            if ($_SESSION['user_role'] == 'admin') {
+            if ($_SESSION['user_role'] == 'Admin') {
                 $the_user_id = $_GET['delete'];
                 $query = "DELETE FROM users WHERE user_id ={$the_user_id}";
                 $delete_user = mysqli_query($connection, $query);
